requirements.txt
pandas==2.3.2
numpy==2.1.3
matplotlib==3.10.5
matplotlib-inline==0.1.7
seaborn==0.13.2
scikit-learn==1.7.1
opencv-python==4.12.0.88
mediapipe==0.10.21
pillow==11.3.0
torch==2.8.0
torchinfo==1.8.0
torchvision==0.23.0
xgboost==3.1.2
lightgbm==4.6.0
catboost==1.2.8
fastapi==0.128.0
pydantic==2.12.5
uvicorn==0.40.0
streamlit==1.52.2
python-multipart==0.0.21